chore(aws): create a models file
